refactor(Twitter): Combined `customize patch` into single class
